+++ ilyvion.LaboratoryMod/Patches/Verse_ScribeLoader_InitLoading_Reverse.cs
@@ -1,8 +1,8 @@
 using System.Reflection;
 using System.Reflection.Emit;
-using Logger = ilvyion.Laboratory.Logger;
+using Logger = ilyvion.Laboratory.Logger;
 
-namespace ilvyion.LaboratoryMod;
+namespace ilyvion.LaboratoryMod;
 
 [HarmonyPatch]
 internal static class Verse_ScribeLoader_InitLoading_Reverse