+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Make it so that the CustomFontManager only patches the font system if it's been enabled. This way it won't affect performance if it goes unused.
 
+### Fixed
+
+-   The name ilyvion was misspelled in several places and file paths as 'ilvyion'
+
 ## [0.1.0] 2024-07-25
 
 ### Added